Update auth methods to take u32 instead of Domain
